Rename module code
