+++ tests/serialize/paulicode.rs
@@ -3,7 +3,7 @@ use std::{
     io::BufReader,
 };
 
-use f2q::codes::qubits::{
+use f2q::code::qubits::{
     Pauli,
     PauliCode,
 };
