+++ tests/serialize/paulisum.rs
@@ -1,5 +1,5 @@
 use f2q::{
-    codes::qubits::{
+    code::qubits::{
         Pauli,
         PauliCode,
         PauliSum,
